+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `--programs-list` argument
 ### Changed
 - `japanese-green-tea` program bug fixed
+- `README.md` updated
 ## [0.9] - 2023-10-04
 ### Added
 - 4 new faces